\name{cal.quakes}
\alias{cal.quakes}

\title{Northern California Earthquake Occurrences}

\usage{data(cal.quakes)}

\format{
  A vector of 849 earthquake (magnitude > 3.0) event times occurring
  between 1/1/1970 and 12/31/2002
}

\description{
  This dataset was taken from the Northern Californa Earthquake Data
  Center at \url{http://quake.geo.berkeley.edu}.  The units of the
  event times are in days since 1/1/1970.  The events took place between
  -121.5 and -121.0 longitude and 36.5 and 37.0 latitude.  
}

\keyword{datasets}
