\name{coerce}
\alias{coerce}
\alias{as.ptprocInit}
\alias{as.ptprocSim}

\title{Coercion between point process objects}
\description{
  Functions for coercion between different class of point process
  objects
}
\usage{
as.ptprocInit(x, ...)
as.ptprocSim(x, ...)
}

\arguments{
  \item{x}{a point process object}
  \item{\dots}{possibly other arguments}
}

\value{
  An object of class \code{"ptprocInit"} or \code{"ptprocSim"},
  respectively.
}

\author{Roger D. Peng}

\keyword{models}
