\name{fires}
\alias{fires}

\title{Northern Los Angeles County Fires}

\usage{data(fires)}

\format{
  A data frame with 313 observations with day of occurrence, x and
  y coordinates.
}


\description{
  Wildfire occurrences in Northern Los Angeles County, California
  between 1976 and 2000.  The spatial units are in scaled feet, taken
  from the NAD 83 state-plane coordinate system.  One unit is equivalent
  to 100,000 feet or 18.9 miles.  The times for the points were produced
  by the \code{date} package and represent the number of days since
  January 1, 1960.
}

\keyword{datasets}
