\name{ptproc.internal}

\alias{ptproc.internal}

\alias{make.optim.logLik}


\title{
  Internal ptproc functions
}

\description{
  Internal functions for \code{ptproc} package.
}

\usage{
make.optim.logLik(ppobj, alpha)
}

\details{
  These functions are not to be called by the user.
}

\keyword{internal}

\author{Roger D. Peng}


